Split intrinsics to test link
