+++ accel-core/src/lib.rs
@@ -1,5 +1,7 @@
 #![feature(platform_intrinsics)]
 
+mod intrinsics;
+pub use intrinsics::*;
 
 pub struct Dim3 {
     pub x: i32,
@@ -63,22 +65,6 @@ impl Idx3 {
     pub fn into_id(&self, dim: Dim3) -> i32 {
         self.x + self.y * dim.x + self.z * dim.x * dim.y
     }
-}
-
-extern "platform-intrinsic" {
-    pub fn nvptx_block_dim_x() -> i32;
-    pub fn nvptx_block_dim_y() -> i32;
-    pub fn nvptx_block_dim_z() -> i32;
-    pub fn nvptx_block_idx_x() -> i32;
-    pub fn nvptx_block_idx_y() -> i32;
-    pub fn nvptx_block_idx_z() -> i32;
-    pub fn nvptx_grid_dim_x() -> i32;
-    pub fn nvptx_grid_dim_y() -> i32;
-    pub fn nvptx_grid_dim_z() -> i32;
-    pub fn nvptx_syncthreads() -> ();
-    pub fn nvptx_thread_idx_x() -> i32;
-    pub fn nvptx_thread_idx_y() -> i32;
-    pub fn nvptx_thread_idx_z() -> i32;
 }
 
 pub fn index() -> isize {
